+++ src/index.ts
@@ -11,7 +11,7 @@ import {
   claimHandle712,
   mockCaip122,
 } from "./signature-requests.js";
-import { GatewayFetchError } from "./error-types.js";
+import { getGatewayAccount } from "./helpers/gateway.js";
 
 type Address = string;
 
@@ -108,38 +108,6 @@ let mockCreationGatewayAccount: AccountResponse = {
 
 export function setMockForCreationGatewayAccount(mockValue: AccountResponse) {
   mockCreationGatewayAccount = mockValue;
-}
-
-/**
- * Fetches a user's account information (if present) from Gateway Services
- *
- * @param gatewayFetchFn Callback for performing request to gateway services
- * @param userAddress ???
- * @returns An 'account response' when the user's account exists, and `null` otherwise
- * @throws `GatewayFetchError` when the request fails
- */
-// TODO: Find a way to unit test without `export`-ing ..?
-export async function getGatewayAccount(
-  gatewayFetchFn: GatewayFetchFn,
-  userAddress: string,
-): Promise<AccountResponse | null> {
-  const response = await gatewayFetchFn(
-    "GET",
-    `/v1/accounts/account/${userAddress}`,
-  );
-
-  if (response.ok) {
-    const body = await response.json() as AccountResponse;
-    // TODO(?): validate with `zod` ..?
-    return body;
-  } else {
-    switch (response.status) {
-      case 404:
-        return null; // The user does not (yet) exist on chain
-      default:
-        throw new GatewayFetchError("Failed GatewayFetchFn for GET Account", response);
-    }
-  }
 }
 
 async function postGatewaySiwf(
